+++ packages/button/stories/Button.stories.js
@@ -2,6 +2,7 @@
 import { withKnobs, text } from "@storybook/addon-knobs";
 import { Button } from "..";
 import ButtonGroupStory01 from "./ButtonGroupStory01.svelte";
+import ButtonLinkStory from "./ButtonLinkStory.svelte";
 import knobs from "./knobs";
 
 export default {
@@ -19,6 +20,13 @@ export const BasicButton = () => ({
 
 export const BasicButtonGroup = () => ({
   Component: ButtonGroupStory01,
+  props: {
+    ...knobs(),
+  },
+});
+
+export const ButtonLink = () => ({
+  Component: ButtonLinkStory,
   props: {
     ...knobs(),
   },